Rename back to "mappings" to appease clippy
