+++ src/encodings/mappings.rs
